requirements.txt
streamlit
pytesseract
pillow
googletrans==4.0.0rc1